add top10 failing test cases
